move PillButton to components package
