+++ app/packages/components/src/components/PillButton/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./PillButton";
