+++ app/packages/components/src/components/index.ts
@@ -17,3 +17,4 @@ export { default as TabOption } from "./TabOption";
 export { default as ThemeProvider, useTheme } from "./ThemeProvider";
 export { default as Selection } from "./Selection";
 export { default as Tooltip } from "./Tooltip";
+export { default as PillButton } from "./PillButton";
