+++ app/packages/core/src/components/utils.tsx
@@ -1,8 +1,7 @@
-import { Tooltip, useTheme } from "@fiftyone/components";
+import { useTheme } from "@fiftyone/components";
 import React, { useState } from "react";
 import styled from "styled-components";
 import { animated, useSpring, useSprings } from "@react-spring/web";
-import { KeyboardArrowUp, KeyboardArrowDown } from "@mui/icons-material";
 
 export const Box = styled.div`
   padding: 1em;
@@ -69,88 +68,6 @@ export const ContentHeader = styled.div`
   display: flex;
   padding-bottom: 0.5rem;
 `;
-
-const PillButtonDiv = animated(styled.div`
-  line-height: 1.5rem;
-  padding: 0.25rem 0.75rem;
-  cursor: pointer;
-  background-color: ${({ theme }) => theme.divider};
-  border-radius: 1rem;
-  border: none;
-  font-weight: bold;
-  display: flex;
-  justify-content: space-between;
-  opacity: 1;
-
-  & > span {
-    text-align: center;
-    margin: 0 0.25rem;
-  }
-  & > svg {
-    display: inline-block;
-    height: 100%;
-  }
-`);
-
-type PillButton = {
-  onClick: (event: Event) => void;
-  open: boolean;
-  highlight: boolean;
-  text?: string;
-  icon?: any;
-  arrow?: boolean;
-  style?: React.CSSProperties;
-  title: string;
-};
-
-export const PillButton = React.memo(
-  React.forwardRef(
-    (
-      {
-        onClick,
-        open,
-        text,
-        icon,
-        highlight,
-        arrow = false,
-        style,
-        title,
-      }: PillButton,
-      ref
-    ) => {
-      const theme = useTheme();
-      const props = useSpring({
-        backgroundColor: !highlight
-          ? theme.background.button
-          : theme.primary.plainColor,
-        color: !highlight ? theme.text.secondary : theme.text.buttonHighlight,
-      });
-
-      const children = (
-        <PillButtonDiv
-          onClick={(e) => {
-            onClick(e);
-          }}
-          onMouseDown={(e) => {
-            e.stopPropagation();
-          }}
-          ref={ref}
-          style={{ ...props, ...style }}
-          title={title}
-        >
-          {text && <span>{text}</span>}
-          {icon}
-          {arrow && (open ? <KeyboardArrowUp /> : <KeyboardArrowDown />)}
-        </PillButtonDiv>
-      );
-      return title ? (
-        <Tooltip text={title}>{children}</Tooltip>
-      ) : (
-        <>{children}</>
-      );
-    }
-  )
-);
 
 export const PopoutDiv = animated(styled.div`
   background-color: ${({ theme }) => theme.background.level2};
